+++ fairMLHealth/__init__.py
@@ -1,2 +1,2 @@
 
-__all__ = ['utils', 'tutorial_and_examples']
+__all__ = ['tools', 'tutorial_and_examples']
