+++ fairMLHealth/tools/format_mimic_data.py
